Create Simple Text Editor
You are given an empty text editor with a string, S, that initially contains an empty string. The editor supports the following operations:

Append: Add a string w at the end of S.
Delete: Delete the last k characters of S.
Print: Print the kth character of S.
Undo: Undo the last not previously undone operation (regardless of type) and return S to the state it was in prior to that operation.
You will be provided q operations to perform, and you must process them in the order given.

Input Format
The first line contains an integer, q, the number of operations.
The next q lines describe the operations:
1 w: Append the string w to S.
2 k: Delete the last k characters of S.
3 k: Print the kth character of S.
4: Undo the last operation.
Constraints
1≤𝑞≤105
1≤q≤10 5
 
1≤∣𝑤∣≤105
1≤∣w∣≤105
(length of w)
1≤𝑘≤∣𝑆∣
1≤k≤∣S∣ (for operations 2 and 3)
The sum of all |w| in one test file will not exceed 
106
106
 .
Output Format
For each operation of type 3, print the kth character of S on a new line.